Move dir detection to soundcraft.dirs module
Just move the directory detection code to a new `soundcraft.dirs`
module, without any significant changes.

This allows us to use the same directory detection code from both
the `soundcraft.setup` and the `soundcraft.gui` modules, so we can
later have the GUI use one of the XDG desktop icon files as its icon
even if the setup code has been removed for a distribution package.
