+++ soundcraft/constants.py
@@ -55,6 +55,7 @@ MGRPATH = "/soundcraft/utils/notepad"
 
 # Helpful for removing remnants of soundcraft-utils 0.4.0 and earlier
 OLD_BASE_EXE_SERVICE = "soundcraft_dbus_service"
+OLD_STATEDIR = "/var/lib/soundcraft-utils"
 
 
 # USB Device vendor and product IDs
